CinemaBot/requirements.txt
aiogram==2.4
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.7.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
git+https://github.com/alberanid/imdbpy
lxml==4.4.2
multidict==4.7.2
pytz==2019.3
SQLAlchemy==1.3.12
typing-extensions==3.7.4.1
yarl==1.4.2
googletrans
flask
gunicorn==20.0.4
kinopoiskpy==0.8.2
python-dateutil==2.6.1
urllib3==1.22
git+git://github.com/nficano/pytube#egg=pytube
